Additional fixes & test clean up
