+++ bundles/core/org.eclipse.smarthome.core.thing/src/main/java/org/eclipse/smarthome/core/thing/i18n/ChannelI18nUtil.java
@@ -12,7 +12,6 @@
  */
 package org.eclipse.smarthome.core.thing.i18n;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import java.util.function.Function;
@@ -21,11 +20,7 @@ import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
 import org.eclipse.smarthome.core.thing.type.ChannelDefinition;
 import org.eclipse.smarthome.core.thing.type.ChannelType;
-import org.eclipse.smarthome.core.thing.type.ChannelTypeRegistry;
-import org.eclipse.smarthome.core.thing.type.ChannelTypeUID;
 import org.osgi.framework.Bundle;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * A utility service which localises {@link ChannelDefinition}.
@@ -38,34 +33,7 @@ import org.osgi.service.component.annotations.Reference;
  *
  */
 @NonNullByDefault
-@Component
-public class ChannelI18nUtil {
-
-    @NonNullByDefault({})
-    private ChannelTypeI18nLocalizationService channelTypeI18nLocalizationService;
-
-    @NonNullByDefault({})
-    private ChannelTypeRegistry channelTypeRegistry;
-
-    @Reference
-    protected void setChannelTypeI18nLocalizationService(
-            ChannelTypeI18nLocalizationService channelTypeI18nLocalizationService) {
-        this.channelTypeI18nLocalizationService = channelTypeI18nLocalizationService;
-    }
-
-    protected void unsetChannelTypeI18nLocalizationService(
-            ChannelTypeI18nLocalizationService channelTypeI18nLocalizationService) {
-        this.channelTypeI18nLocalizationService = null;
-    }
-
-    @Reference
-    protected void setChannelTypeRegistry(ChannelTypeRegistry channelTypeRegistry) {
-        this.channelTypeRegistry = channelTypeRegistry;
-    }
-
-    protected void unsetChannelTypeRegistry(ChannelTypeRegistry channelTypeRegistry) {
-        this.channelTypeRegistry = null;
-    }
+public interface ChannelI18nUtil {
 
     /**
      * Localise a list of {@link ChannelDefinition}s.
@@ -81,29 +49,5 @@ public class ChannelI18nUtil {
             final List<ChannelDefinition> channelDefinitions,
             final Function<ChannelDefinition, @Nullable String> channelLabelResolver,
             final Function<ChannelDefinition, @Nullable String> channelDescriptionResolver,
-            final @Nullable Locale locale) {
-        List<ChannelDefinition> localizedChannelDefinitions = new ArrayList<>(channelDefinitions.size());
-        for (final ChannelDefinition channelDefinition : channelDefinitions) {
-            String channelLabel = channelLabelResolver.apply(channelDefinition);
-            String channelDescription = channelDescriptionResolver.apply(channelDefinition);
-            if (channelLabel == null || channelDescription == null) {
-                ChannelTypeUID channelTypeUID = channelDefinition.getChannelTypeUID();
-                ChannelType channelType = channelTypeRegistry.getChannelType(channelTypeUID, locale);
-                if (channelType != null) {
-                    ChannelType localizedChannelType = channelTypeI18nLocalizationService
-                            .createLocalizedChannelType(bundle, channelType, locale);
-                    if (channelLabel == null) {
-                        channelLabel = localizedChannelType.getLabel();
-                    }
-                    if (channelDescription == null) {
-                        channelDescription = localizedChannelType.getDescription();
-                    }
-                }
-            }
-            localizedChannelDefinitions
-                    .add(new ChannelDefinition(channelDefinition.getId(), channelDefinition.getChannelTypeUID(),
-                            channelDefinition.getProperties(), channelLabel, channelDescription));
-        }
-        return localizedChannelDefinitions;
-    }
+            final @Nullable Locale locale);
 }
